Update revision 10042019
Fix and clear code
Add menu axis limit to menu advanced

+++ MK4duo/src/lcd/menu/menu_mixer.cpp
@@ -41,7 +41,7 @@
     lcdui.encoder_direction_normal();
     ENCODER_RATE_MULTIPLY(true);
     if (lcdui.encoderPosition != 0) {
-      mixer.gradient.start_z += float((int)lcdui.encoderPosition) * 0.1;
+      mixer.gradient.start_z += float((int16_t)lcdui.encoderPosition) * 0.1;
       lcdui.encoderPosition = 0;
       NOLESS(mixer.gradient.start_z, 0);
       NOMORE(mixer.gradient.start_z, Z_MAX_BED);
@@ -66,7 +66,7 @@
     lcdui.encoder_direction_normal();
     ENCODER_RATE_MULTIPLY(true);
     if (lcdui.encoderPosition != 0) {
-      mixer.gradient.end_z += float((int)lcdui.encoderPosition) * 0.1;
+      mixer.gradient.end_z += float((int16_t)lcdui.encoderPosition) * 0.1;
       lcdui.encoderPosition = 0;
       NOLESS(mixer.gradient.end_z, 0);
       NOMORE(mixer.gradient.end_z, Z_MAX_BED);
@@ -176,7 +176,7 @@ void lcd_mixer_mix_edit() {
   #elif HAS_GRADIENT_MIX
 
     if (lcdui.encoderPosition != 0) {
-      mixer.mix[0] += (int)lcdui.encoderPosition;
+      mixer.mix[0] += (int16_t)lcdui.encoderPosition;
       lcdui.encoderPosition = 0;
       if (mixer.mix[0] < 0) mixer.mix[0] += 101;
       if (mixer.mix[0] > 100) mixer.mix[0] -= 101;
